Prevent Entire Editor Header from re rendering (#6415)
* Seperate save indicator without rerendering entire Editor Header

* stop unnecessary update to Saved message every second

* moving raw strings to messages
