+++ app/client/src/components/ads/Tooltip.tsx
@@ -3,6 +3,7 @@ import { CommonComponentProps } from "./common";
 import { Position, Tooltip, PopperBoundary } from "@blueprintjs/core";
 import { GLOBAL_STYLE_TOOLTIP_CLASSNAME } from "globalStyles/tooltip";
 import { Modifiers } from "popper.js";
+import { noop } from "lodash";
 
 type Variant = "dark" | "light";
 
@@ -21,6 +22,7 @@ type TooltipProps = CommonComponentProps & {
   minimal?: boolean;
   modifiers?: Modifiers;
   isOpen?: boolean;
+  onOpening?: typeof noop;
 };
 
 const portalContainer = document.getElementById("tooltip-root");
@@ -39,6 +41,7 @@ function TooltipComponent(props: TooltipProps) {
         preventOverflow: { enabled: false },
         ...props.modifiers,
       }}
+      onOpening={props.onOpening}
       openOnTargetFocus={props.openOnTargetFocus}
       popoverClassName={GLOBAL_STYLE_TOOLTIP_CLASSNAME}
       portalContainer={portalContainer as HTMLDivElement}
