+++ app/client/src/constants/messages.ts
@@ -361,6 +361,9 @@ export const DELETE_CONFIRMATION_MODAL_TITLE = () => `Are you sure?`;
 export const DELETE_CONFIRMATION_MODAL_SUBTITLE = (name?: string | null) =>
   `You want to remove ${name} from this organization`;
 
+//Editor Page
+export const EDITOR_HEADER_SAVE_INDICATOR = () => "Saved";
+
 // Generate page from DB Messages
 
 export const UNSUPPORTED_PLUGIN_DIALOG_TITLE = () =>
